+++ src/commands/prepare.ts
@@ -28,7 +28,7 @@ import { getDefaultBranch, getGithubClient } from '../utils/githubApi';
 import { isDryRun } from '../utils/helpers';
 import { formatJson } from '../utils/strings';
 import { sleepAsync, spawnProcess } from '../utils/system';
-import { isValidVersion, versionToTag } from '../utils/version';
+import { isValidVersion } from '../utils/version';
 
 import { handler as publishMainHandler, PublishOptions } from './publish';
 
@@ -43,6 +43,12 @@ export const builder: CommandBuilder = (yargs: Argv) =>
   yargs
     .positional('NEW-VERSION', {
       description: 'The new version you want to release',
+      type: 'string',
+    })
+    .option('rev', {
+      alias: 'r',
+      default: null,
+      description: 'Source revision to prepare the release from.',
       type: 'string',
     })
     .option('no-push', {
@@ -68,9 +74,11 @@ export const builder: CommandBuilder = (yargs: Argv) =>
     .check(checkVersionOrPart);
 
 /** Command line options. */
-interface ReleaseOptions {
+interface PrepareOptions {
   /** The new version to release */
   newVersion: string;
+  /** The base revision to release */
+  rev: string;
   /** Do not perform basic git checks */
   noGitChecks: boolean;
   /** Do not check for changelog */
@@ -118,6 +126,7 @@ function checkVersionOrPart(argv: Arguments<any>, _opt: any): any {
  */
 async function createReleaseBranch(
   git: SimpleGit,
+  rev: string,
   newVersion: string,
   releaseBranchPrefix?: string
 ): Promise<string> {
@@ -137,7 +146,7 @@ async function createReleaseBranch(
   }
 
   if (!isDryRun()) {
-    await git.checkoutLocalBranch(branchName);
+    await git.checkoutBranch(branchName, rev);
     logger.info(`Created a new release branch: "${branchName}"`);
     logger.info(`Switched to branch "${branchName}"`);
   } else {
@@ -245,16 +254,7 @@ export async function runPreReleaseCommand(
  * @param defaultBranch Default branch of the remote repository
  * @param checkGitStatus Set to true to enable the check
  */
-async function checkGitState(
-  git: SimpleGit,
-  defaultBranch: string,
-  checkGitStatus = true
-): Promise<void> {
-  if (!checkGitStatus) {
-    logger.warn('Not checking the status of the local repository');
-    return;
-  }
-
+async function checkGitState(git: SimpleGit, rev: string): Promise<void> {
   logger.info('Checking the local repository status...');
   const isRepo = await git.checkIsRepo();
   if (!isRepo) {
@@ -263,14 +263,6 @@ async function checkGitState(
   const repoStatus = await git.status();
   logger.debug('Repository status:', formatJson(repoStatus));
 
-  // Check that we are on master
-  // TODO check what's here when we are in a detached state
-  const currentBranch = repoStatus.current;
-  if (defaultBranch !== currentBranch) {
-    reportError(
-      `Please switch to your default branch (${defaultBranch}) first`
-    );
-  }
   if (
     repoStatus.conflicted.length ||
     repoStatus.created.length ||
@@ -286,10 +278,9 @@ async function checkGitState(
     );
   }
 
-  if (repoStatus.ahead > 0) {
-    reportError(
-      `Your repository has unpushed changes: the current branch is ${repoStatus.ahead} commits ahead.`,
-      logger
+  if (repoStatus.current !== rev) {
+    logger.warn(
+      `You are releasing from '${rev}', not '${repoStatus.current}' which you are currently on.`
     );
   }
 }
@@ -332,30 +323,6 @@ async function execPublish(newVersion: string): Promise<never> {
     throw e;
   }
   throw new Error('Unreachable');
-}
-
-/**
- * Checks that there is no corresponding git tag for the given version
- *
- * @param git Local git client
- * @param newVersion Version we're about to release
- * @param checkGitStatus Set to true to enable the check
- */
-async function checkForExistingTag(
-  git: SimpleGit,
-  newVersion: string,
-  checkGitStatus = true
-): Promise<void> {
-  if (!checkGitStatus) {
-    logger.warn('Not checking if the version (git tag) already exists');
-  }
-
-  const gitTag = versionToTag(newVersion);
-  const existingTags = await git.tags();
-  if (existingTags.all.indexOf(gitTag) > -1) {
-    reportError(`Git tag "${gitTag}" already exists!`);
-  }
-  logger.debug(`Git tag ${gitTag} does not exist yet.`);
 }
 
 /**
@@ -475,7 +442,7 @@ async function switchToDefaultBranch(
  *
  * @param argv Command-line arguments
  */
-export async function releaseMain(argv: ReleaseOptions): Promise<any> {
+export async function prepareMain(argv: PrepareOptions): Promise<any> {
   // Get repo configuration
   const config = getConfiguration();
   const githubConfig = await getGlobalGithubConfig();
@@ -497,12 +464,14 @@ export async function releaseMain(argv: ReleaseOptions): Promise<any> {
     githubConfig.repo
   );
   logger.debug(`Default branch for the repo:`, defaultBranch);
+  const rev = argv.rev || defaultBranch;
 
-  // Check that we're in an acceptable state for the release
-  await checkGitState(git, defaultBranch, !argv.noGitChecks);
-
-  // Check whether the version/tag already exists
-  await checkForExistingTag(git, newVersion, !argv.noGitChecks);
+  if (argv.noGitChecks) {
+    logger.info('Not checking the status of the local repository');
+  } else {
+    // Check that we're in an acceptable state for the release
+    await checkGitState(git, rev);
+  }
 
   // Check & update the changelog
   await prepareChangelog(
@@ -517,6 +486,7 @@ export async function releaseMain(argv: ReleaseOptions): Promise<any> {
   // exists.
   const branchName = await createReleaseBranch(
     git,
+    rev,
     newVersion,
     config.releaseBranchPrefix
   );
@@ -551,7 +521,9 @@ export async function releaseMain(argv: ReleaseOptions): Promise<any> {
     );
   }
 
-  await switchToDefaultBranch(git, defaultBranch);
+  if (!argv.rev) {
+    await switchToDefaultBranch(git, defaultBranch);
+  }
 }
 
 /**
@@ -567,7 +539,7 @@ export const handler = async (args: {
   [argName: string]: any;
 }): Promise<void> => {
   try {
-    return await releaseMain(args as ReleaseOptions);
+    return await prepareMain(args as PrepareOptions);
   } catch (e) {
     handleGlobalError(e);
   }